Added schedulers and readcount()

+++ user/user.h
@@ -24,7 +24,7 @@ int getpid(void);
 char* sys_sbrk(int,int);
 int pause(int);
 int uptime(void);
-int getreadcount(void);
+uint64 getreadcount(void);
 
 // ulib.c
 int stat(const char*, struct stat*);
